Fix bug related to nested groupings
Prior to this commit if a nested grouping was encountered (ie.
`((this && that) || the_other)` the gem would raise the following error:

```
NoMethodError:
  undefined method `strip' for nil:NilClass
```

It appears that in this scenario it's possible that the `operator`
variable hasn't yet been set. The proposed fix is simply coercing the
variable into a `String` before attempting to call the `#strip`
function.